requirements.txt
dj-database-url==0.5.0
Django==3.0.8
django-bootstrap4==2.2.0
django-cleanup==4.0.0
django-crispy-forms==1.8.1
django-heroku==0.3.1
django-mysql==3.3.0
gunicorn==20.0.4
mysqlclient==1.4.6
newsapi-python==0.2.6
psycopg2==2.8.5
requests==2.23.0
requests-file==1.5.1
